phase 2 fixes
